+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Add `createServiceDataSetKey` function to easy create a `DataSetKey` with `Instance`
+- Add `ResourceAvailability` and `Audience` metric module
 
 ## 1.0.0 - 2019-01-31
 - Initial implementation